+++ test/run_all_versions.sh
@@ -252,6 +252,7 @@ SUITES=(
 	planner_choice_quality
 	preimage_rewrite
 	projection_drop_column
+	projection_parallel
 	projection_privilege
 	projection_rename_restore
 	projection_rewrite